requirements.txt
fastapi[all]
uvicorn
python-multipart
pytesseract
pillow
openai
python-dotenv
alembic
sqlalchemy
psycopg2-binary
pdf2image
xero-python>=4.1.0
requests